doc: update README and add Makefiles
