+++ client/Makefile
@@ -1,0 +1,5 @@
+PORT = 8080
+
+run:
+	@npm install
+	@npm run dev -- --port $(PORT)
